Reorganized components code.
