+++ src/App.js
@@ -12,7 +12,7 @@ import store from "./redux/store";
 import { SET_AUTHENTICATED } from "./redux/types";
 import { logoutUser, getUserData } from "./redux/actions/userActions";
 // Components
-import Navbar from "./components/Navbar";
+import Navbar from "./components/Layout/Navbar";
 import AuthRoute from "./util/AuthRoute";
 // Pages
 import home from "./pages/home";
