+++ src/components/Scream/LikeButton.js
@@ -2,13 +2,13 @@ import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
 // Components
-import MyButton from "../util/MyButton";
+import MyButton from "../../util/MyButton";
 // Icons
 import FavoriteIcon from "@material-ui/icons/Favorite";
 import FavoriteBorder from "@material-ui/icons/FavoriteBorder";
 // Redux
 import { connect } from "react-redux";
-import { likeScream, unlikeScream } from "../redux/actions/dataActions";
+import { likeScream, unlikeScream } from "../../redux/actions/dataActions";
 
 class LikeButton extends Component {
   likedScream = () => {
