+++ src/components/Scream/Scream.js
@@ -5,7 +5,7 @@ import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import PropTypes from "prop-types";
 // Components
-import MyButton from "../util/MyButton";
+import MyButton from "../../util/MyButton";
 import DeleteScream from "./DeleteScream";
 import LikeButton from "./LikeButton";
 import ScreamDialog from "./ScreamDialog";
