+++ src/pages/home.js
@@ -3,8 +3,8 @@ import Grid from "@material-ui/core/Grid";
 import PropTypes from "prop-types";
 
 // Components
-import Scream from "../components/Scream";
-import Profile from "../components/Profile";
+import Scream from "../components/Scream/Scream";
+import Profile from "../components/Profile/Profile";
 
 import { connect } from "react-redux";
 import { getScreams } from "../redux/actions/dataActions";
